move component in motion template to component/motion folder
